glm branch: double -> float in player/.
svn path=/branches/libavg_glm/; revision=6910

+++ src/player/AppleTrackpadInputDevice.cpp
@@ -61,7 +61,7 @@ void AppleTrackpadInputDevice::start()
 }
 
 void AppleTrackpadInputDevice::onData(int device, Finger* pFingers, int numFingers, 
-        double timestamp, int frame)
+        float timestamp, int frame)
 {
     boost::mutex::scoped_lock lock(getMutex());
     for (int i = 0; i < numFingers; i++) {
@@ -99,7 +99,7 @@ TouchEventPtr AppleTrackpadInputDevice::createEvent(int avgID, Finger* pFinger,
     glm::vec2 size = getWindowSize();
     IntPoint pos(pFinger->normalized.pos.x*size.x, (1-pFinger->normalized.pos.y)*size.y);
     glm::vec2 speed(pFinger->normalized.vel.x*size.x, pFinger->normalized.vel.y*size.y);
-    double eccentricity = pFinger->majorAxis/pFinger->minorAxis;
+    float eccentricity = pFinger->majorAxis/pFinger->minorAxis;
     glm::vec2 majorAxis = fromPolar(pFinger->angle, pFinger->majorAxis);
     majorAxis.y = -majorAxis.y;
     glm::vec2 minorAxis = fromPolar(pFinger->angle+1.57, pFinger->minorAxis);
