+++ src/player/AppleTrackpadInputDevice.h
@@ -66,8 +66,8 @@ public:
     virtual void start();
 
 private:
-    void onData(int device, Finger *data, int nFingers, double timestamp, int frame);
-    static int callback(int device, Finger *data, int nFingers, double timestamp, 
+    void onData(int device, Finger *data, int nFingers, float timestamp, int frame);
+    static int callback(int device, Finger *data, int nFingers, float timestamp, 
             int frame);
     TouchEventPtr createEvent(int avgID, Finger* pFinger, Event::Type eventType);
